+move +add fake enum
