--- Ability/Test.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: b2b22e2dcb18467c9f6caa33e3c946a9
-timeCreated: 1714717912
